src/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
pandas==2.3.2
sqlalchemy==2.0.43
python-dotenv==1.1.1
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
python-multipart

